Problem 74 - Search a 2D Matrix
This solution passed all of the LeetCode tests, running faster
than 93.31% of all Python 3 submissions, but using less memory
than only 8.89% of all Python 3 submissions.

The reason for the high speed and memory usage is likely due to
the recursive implementation of the binary search algorithm, which
requires stack frame allocation for each iteration of the algorithm,
while imperative implementations do not.

One notable point to make is that the initial submission was wrong,
since I was checking only whether the last element in the current
matrix row was greater than the target value, when in reality the
correct method is to check whether the last element of the current
row is greater than or equal to the target value.