Problem 68 - Text Justification
This problem originally failed because I forgot to check
whether a word's length was equal to the maximum width
allowed. I was originally checking to see whether the
length of the word and the length of the current line
were greater than the maximum width.

This was actually two errors, since I was also forgetting
to account for the necessary space required when adding a
word to the line.